refactor: DealerTest 에 있던 Result 관련 테스트 를 ResultTest 로 이동
